+++ dorm_annotations/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* Add `UidType` class, that handles unique-identification on models
+* Model now implements schema
+
 ## 0.0.1
 
 * TODO: Describe initial release.